QPID-7753: Correct return type of getAllocatedDirectMemorySize from int to long

+++ broker-core/src/main/java/org/apache/qpid/server/bytebuffer/QpidByteBuffer.java
@@ -857,7 +857,7 @@ public class QpidByteBuffer implements AutoCloseable
         return _pooledBufferSize;
     }
 
-    public static int getAllocatedDirectMemorySize()
+    public static long getAllocatedDirectMemorySize()
     {
         return _pooledBufferSize * getNumberOfActivePooledBuffers();
     }
